Refactor image compression out of StripeAPI
